add GetN method to retrieve batches of available items (#11)
* add GetN method to retrieve batches of available items

GetN retrieves up to n items that are currently available in the queue. If there are no items currently available, then none are returned and GetN does
not wait for any.

GetN is used to poll the DSQueue for items and return batches of those items. This is the most efficient way of fetching currently available items.

GetN and Out can both be used to read items from the DSQueue, but they should not be used concurrently as items will be returned by one or tha other indeterminately.

* improve tests